Several changes
Modfies Board's checkWord method, and added methods containsOnlyOneTile, checkIfConnectedWordIsValid, and letterTouchesAnotherWord to implement the following Scrabble rules:

New words may be formed by:
Adding one or more letters to a word or letters already on the board.

Placing a word at right angles to a word already on the board. The new word must use one of the letters already on the board or must add a letter to it.

Placing a complete word parallel to a word already played so that adjacent letters also form complete words.

Note: The program works if and only if the first move starts at the centre square. That issue must be fixed for the program to work at all times.